@@ -33,25 +33,27 @@ package edu.iu.uits.lms.gct.services;
  * #L%
  */
 
-import edu.iu.uits.lms.gct.config.ToolConfig;
+import edu.iu.uits.lms.gct.config.SecurityConfig;
 import edu.iu.uits.lms.gct.controller.rest.PropertiesRestController;
 import edu.iu.uits.lms.gct.repository.GctPropertyRepository;
 import edu.iu.uits.lms.lti.config.TestUtils;
+import edu.iu.uits.lms.lti.repository.DefaultInstructorRoleRepository;
 import io.jsonwebtoken.SignatureAlgorithm;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 
 import java.time.Instant;
@@ -63,7 +65,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(controllers = {PropertiesRestController.class}, properties = {"oauth.tokenprovider.url=http://foo"})
-@Import(ToolConfig.class)
+@ContextConfiguration(classes = {PropertiesRestController.class, SecurityConfig.class})
 @Slf4j
 @ActiveProfiles("none")
 public class RestLaunchSecurityTest {
@@ -73,9 +75,15 @@ public class RestLaunchSecurityTest {
    @MockBean
    private GctPropertyRepository gctPropertyRepository;
 
+   @MockBean
+   private DefaultInstructorRoleRepository defaultInstructorRoleRepository;
+
+   @MockBean
+   private ClientRegistrationRepository clientRegistrationRepository;
+
    @Test
    public void restNoAuthnLaunch() throws Exception {
-      //This is a secured endpoint and should not not allow access without authn
+      //This is a secured endpoint and should not allow access without authn
       SecurityContextHolder.getContext().setAuthentication(null);
       mvc.perform(get("/rest/properties/test/foo")
             .header(HttpHeaders.USER_AGENT, TestUtils.defaultUseragent())
@@ -90,7 +98,7 @@ public class RestLaunchSecurityTest {
       Collection<GrantedAuthority> authorities = AuthorityUtils.createAuthorityList("SCOPE_lms:rest", "ROLE_LMS_REST_ADMINS");
       JwtAuthenticationToken token = new JwtAuthenticationToken(jwt, authorities);
 
-      //This is a secured endpoint and should not not allow access without authn
+      //This is a secured endpoint and should not allow access without authn
       mvc.perform(get("/rest/properties/test/foo")
             .header(HttpHeaders.USER_AGENT, TestUtils.defaultUseragent())
             .contentType(MediaType.APPLICATION_JSON)
@@ -105,7 +113,7 @@ public class RestLaunchSecurityTest {
       Collection<GrantedAuthority> authorities = AuthorityUtils.createAuthorityList("SCOPE_read", "ROLE_NONE_YA");
       JwtAuthenticationToken token = new JwtAuthenticationToken(jwt, authorities);
 
-      //This is a secured endpoint and should not not allow access without authn
+      //This is a secured endpoint and should not allow access without authn
       mvc.perform(get("/rest/properties/test/foo")
             .header(HttpHeaders.USER_AGENT, TestUtils.defaultUseragent())
             .contentType(MediaType.APPLICATION_JSON)
